+++ src/pages/Scene.tsx
@@ -1,0 +1,5 @@
+import SceneGame from '../ui/SceneGame';
+
+export default function Scene() {
+  return <SceneGame />;
+}